Create Cobalt Strike Detection
This query looks for network connections to known Cobalt Strike command-and-control (C2) servers by searching for events in the SecurityNetworkCommunication table where the remote IP address matches a list of known C2 IPs. It then filters the results to only include connections on ports 80 and 443, which are commonly used by Cobalt Strike. The query then extracts relevant fields and joins the network connections to the SecurityEvent table to identify the processes making the connections. Finally, it outputs the results and includes the name of the rule in the output.